+++ Changelog.md
@@ -1,6 +1,7 @@
 # [Beta 1.1 (Unreleased)]()
 
 - Reimplements eye tracking
+- Fix stereo IPD mismatch between server render and client warp (tracker used 64mm, warp used 65mm)
 
 # [Beta 1.0](https://github.com/openlyst/piconeo2-wivrn/releases/download/Beta-1.0/wivrn-pvr-Beta-1.0.apk)